Доработки после merge
