+++ src/main/java/faang/school/achievement/handler/mentorship/SenseiAchievementHandler.java
@@ -1,8 +1,8 @@
 package faang.school.achievement.handler.mentorship;
 
-import faang.school.achievement.service.achievementprogress.DefaultAchievementProgressServiceService;
-import faang.school.achievement.service.cache.DefaultAchievementCacheService;
-import faang.school.achievement.service.userachievement.DefaultUserAchievementService;
+import faang.school.achievement.service.achievementprogress.AchievementProgressService;
+import faang.school.achievement.service.cache.AchievementCacheService;
+import faang.school.achievement.service.userachievement.UserAchievementService;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -11,9 +11,9 @@ public class SenseiAchievementHandler extends MentorshipEventHandler {
     private final static String SENSEI = "sensei";
 
     public SenseiAchievementHandler(
-            DefaultAchievementProgressServiceService achievementProgressService,
-            DefaultAchievementCacheService achievementCacheService,
-            DefaultUserAchievementService userAchievementService) {
+            AchievementProgressService achievementProgressService,
+            AchievementCacheService achievementCacheService,
+            UserAchievementService userAchievementService) {
         super(achievementProgressService, achievementCacheService, userAchievementService);
     }
