+++ src/main/java/faang/school/achievement/service/achievementprogress/DefaultAchievementProgressService.java
@@ -3,9 +3,15 @@ package faang.school.achievement.service.achievementprogress;
 import faang.school.achievement.exception.AchievementProgressNotFoundException;
 import faang.school.achievement.model.AchievementProgress;
 import faang.school.achievement.repository.AchievementProgressRepository;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.persistence.OptimisticLockException;
+import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,6 +27,9 @@ public class DefaultAchievementProgressService implements AchievementProgressSer
 
     private final AchievementProgressRepository achievementProgressRepository;
 
+    @PersistenceContext
+    private EntityManager entityManager;
+
     @Transactional
     @Override
     public void createProgressIfNecessary(long userId, long achievementId) {
@@ -34,5 +43,27 @@ public class DefaultAchievementProgressService implements AchievementProgressSer
         return achievementProgressRepository
                 .findByUserIdAndAchievementId(userId, achievementId)
                 .orElseThrow(() -> new AchievementProgressNotFoundException("Achievement progress not found"));
+    }
+
+    @Transactional
+    @Retryable(
+            value = {OptimisticLockException.class},
+            maxAttempts = 5,
+            backoff = @Backoff(delay = 100)
+    )
+    @Override
+    public AchievementProgress progressIncrement(long id) {
+        log.info("Starting progressIncrement for id: {}", id);
+        AchievementProgress progress = entityManager.find(AchievementProgress.class, id);
+
+        if (progress == null) {
+            throw new EntityNotFoundException(String.format("Achievement progress with id %d not found", id));
+        }
+
+        progress.increment();
+
+        log.info("Progress incremented for id: {}", id);
+
+        return progress;
     }
 }
