+++ src/main/java/faang/school/achievement/service/userachievement/DefaultUserAchievementService.java
@@ -1,5 +1,6 @@
 package faang.school.achievement.service.userachievement;
 
+import faang.school.achievement.model.Achievement;
 import faang.school.achievement.model.UserAchievement;
 import faang.school.achievement.repository.UserAchievementRepository;
 import faang.school.achievement.service.achievement.AchievementService;
@@ -30,15 +31,16 @@ public class DefaultUserAchievementService implements UserAchievementService {
 
     @Transactional
     @Override
-    public void assignAchievementToUser(long userId, long achievementId) { //TODO возможно стоит объединить методы
+    public void assignAchievementToUser(long userId, long achievementId) {
         UserAchievement userAchievement = UserAchievement.builder()
-                .achievement(achievementService.getAchievementById(achievementId))
+                .achievement(achievementService.getAchievement(achievementId))
                 .userId(userId)
                 .build();
 
         userAchievementRepository.save(userAchievement);
     }
 
+    @Override
     public void giveAchievement(long userId, Achievement achievement) {
         String title = achievement.getTitle();
         log.info("Starting giving \"{}\" achievement for user {}...", title, userId);
